requirements.txt
git+https://github.com/codezjx/netease-cloud-music-dl
python-telegram-bot

socksio